+++ internal/model/cart.go
@@ -6,9 +6,9 @@ import (
 
 // CartItem represents an item in the cart
 type CartItem struct {
-	ID        uint      `json:"id" gorm:"column:id;primaryKey"`
-	UserID    uint      `json:"userID" gorm:"column:user_id;index;not null"`
-	ProductID uint      `json:"productID" gorm:"column:product_id;index;not null"`
+	ID        uint64    `json:"id" gorm:"column:id;primaryKey"`
+	UserID    uint64    `json:"userID" gorm:"column:user_id;index;not null"`
+	ProductID uint64    `json:"productID" gorm:"column:product_id;index;not null"`
 	Quantity  int       `json:"quantity" gorm:"default:1"`
 	Selected  bool      `json:"selected" gorm:"default:true"`
 	CreatedAt time.Time `json:"createdAt" gorm:"column:created_at"`
@@ -18,8 +18,8 @@ type CartItem struct {
 
 // CartItemResponse is the response for cart item list
 type CartItemResponse struct {
-	ID         uint    `json:"id"`
-	ProductID  uint    `json:"product_id"`
+	ID         uint64  `json:"id"`
+	ProductID  uint64  `json:"product_id"`
 	Quantity   int     `json:"quantity"`
 	Selected   bool    `json:"selected"`
 	Name       string  `json:"name"`
