+++ internal/model/order.go
@@ -19,14 +19,14 @@ const (
 
 // Order represents an order
 type Order struct {
-	ID            uint        `json:"id" gorm:"column:id;primaryKey"`
-	UserID        uint        `json:"userID" gorm:"column:user_id;index;not null"`
+	ID            uint64      `json:"id" gorm:"column:id;primaryKey"`
+	UserID        uint64      `json:"userID" gorm:"column:user_id;index;not null"`
 	OrderNo       string      `json:"orderNo" gorm:"column:order_no;uniqueIndex;not null"`
 	TotalAmount   float64     `json:"totalAmount" gorm:"column:total_amount;type:decimal(10,2);not null"`
 	PaymentAmount float64     `json:"paymentAmount" gorm:"column:payment_amount;type:decimal(10,2);not null"`
 	Status        int         `json:"status" gorm:"column:status;default:0"`
 	PaymentTime   time.Time   `json:"paymentTime" gorm:"column:payment_time"`
-	AddressID     uint        `json:"addressID" gorm:"column:address_id"`
+	AddressID     uint64      `json:"addressID" gorm:"column:address_id"`
 	ReceiverName  string      `json:"receiverName" gorm:"column:receiver_name"`
 	ReceiverPhone string      `json:"receiverPhone" gorm:"column:receiver_phone"`
 	Address       string      `json:"address" gorm:"column:address"`
@@ -38,9 +38,9 @@ type Order struct {
 
 // OrderItem represents an item in an order
 type OrderItem struct {
-	ID        uint      `json:"id" gorm:"column:id;primaryKey"`
-	OrderID   uint      `json:"orderID" gorm:"column:order_id;index;not null"`
-	ProductID uint      `json:"productID" gorm:"column:product_id;index;not null"`
+	ID        uint64    `json:"id" gorm:"column:id;primaryKey"`
+	OrderID   uint64    `json:"orderID" gorm:"column:order_id;index;not null"`
+	ProductID uint64    `json:"productID" gorm:"column:product_id;index;not null"`
 	Quantity  int       `json:"quantity" gorm:"column:quantity;not null"`
 	Price     float64   `json:"price" gorm:"column:price;type:decimal(10,2);not null"`
 	Name      string    `json:"name" gorm:"column:name;not null"`
@@ -51,10 +51,10 @@ type OrderItem struct {
 
 // OrderRequest represents the order creation request
 type OrderRequest struct {
-	AddressID uint   `json:"address_id" binding:"required"`
-	CartIDs   []uint `json:"cart_ids"`                 // Optional for cart checkout
-	ProductID uint   `json:"product_id"`               // For direct purchase
-	Quantity  int    `json:"quantity" binding:"min=1"` // For direct purchase
+	AddressID uint64   `json:"address_id" binding:"required"`
+	CartIDs   []uint64 `json:"cart_ids"`                 // Optional for cart checkout
+	ProductID uint64   `json:"product_id"`               // For direct purchase
+	Quantity  int      `json:"quantity" binding:"min=1"` // For direct purchase
 }
 
 // PaymentResponse represents the payment response
